refactor: 프로젝트 관리 페이지 내 컴포넌트 파일 분리
